Refactor controllers: move ResourceNotFoundException import to utils package
