+++ server/src/main/java/com/kalyvianakis/estiator/io/controller/UserController.java
@@ -2,7 +2,7 @@ package com.kalyvianakis.estiator.io.controller;
 
 import com.kalyvianakis.estiator.io.component.patcher.UserPatcher;
 import com.kalyvianakis.estiator.io.enums.ScheduleStatus;
-import com.kalyvianakis.estiator.io.global.ResourceNotFoundException;
+import com.kalyvianakis.estiator.io.utils.ResourceNotFoundException;
 import com.kalyvianakis.estiator.io.model.MessageResponse;
 import com.kalyvianakis.estiator.io.model.Schedule;
 import com.kalyvianakis.estiator.io.model.ScheduleRequest;
